Library upgrade
  * introduce navigation filter
  * anise management improvements
  * introduce Timescale transpositions
