+++ Cargo.toml
@@ -11,13 +11,28 @@ categories = ["science", "science::geo"]
 edition = "2021"
 
 [features]
-default = ["sp3"]
+default = ["flate2", "sp3", "navigation"]
+
+# Support Gzip compressed files natively
+flate2 = [
+    "rinex/flate2",
+    "sp3?/flate2",
+]
 
 # Unlock support of high precision SP3 files.
 # When targetting highest precision analysis and solutions, like in PPP,
 # SP3 files are mandatory. When deactivated, we simply cannot load
 # such files into a context, only RINEX post processing is possible.
 sp3 = ["dep:sp3"]
+
+# Unlock Post Processed Navigation.
+# Without this option, you are limited to context setup & analysis (basic processing).
+# This option unlocks most advanced solvers.
+navigation = [
+    "dep:anise",
+    "rinex/nav",
+    "sp3?/anise",
+]
 
 # Unlock "embedded ephemeris" option,
 # which removes the need of Internet access on very first (ever) deployment
@@ -42,17 +57,17 @@ strum_macros = "0.26"
 itertools = "0.14.0"
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 
-anise = "0.5.3"
 hifitime = "4.0"
+anise = { version = "0.5.3", optional = true }
 
 gnss-rs = { version = "2.3.5", features = ["serde"] }
-gnss-qc-traits = { version = "0.1.1", features = ["processing"] }
+gnss-qc-traits = { path = "../qc-traits", features = ["processing"] }
 
 maud = "0.26"
 plotly = "0.12"
 
-sp3 = { version = "1.1.2", features = ["qc", "processing", "anise", "serde"], optional = true }
-rinex = { git = "https://github.com/rtk-rs/rinex", branch = "main", features = ["qc", "processing", "obs", "clock", "nav", "ionex", "serde"]  }
+sp3 = { path = "../sp3", features = ["qc", "processing", "serde"], optional = true }
+rinex = { path = "../rinex", features = ["qc", "processing", "obs", "clock", "ionex", "serde"]  }
 
 [dev-dependencies]
 serde_json = "1"
