+++ README.md
@@ -8,33 +8,43 @@ GNSS Quality Control
 
 [![License](https://img.shields.io/badge/license-MPL_2.0-orange?style=for-the-badge&logo=mozilla)](https://github.com/rtk-rs/qc-traits/blob/main/LICENSE)
 
-The Quality Control traits library (`gnss-qc-traits`) is a small library
-that offers the basic operations to form a geodesy processing pipeline, as used by
-our [Quality Control library](https://github.com/rtk-rs/gnss-qc).
+The GNSS Quality Control (QC) library is an advanced library that proposes
+from basic to advanced GNSS and Geodesy processing pipelines.
 
+It is made possible by the complex combination of several frameworks and libraries.
+It is important to understand this library's features & options.
 
-The `GNSS-Qc` (Quality Control) library answers the demanding task
-of GNSS data (post) processing. This topic usually involves several different formats
-at the input of a processing pipeline. 
-
-A pipeline may be formed using all types that implement
-our [Qc Traits](https://github.com/rtk-rs/qc-traits) (low level traits).
-
-A processing pipeline usually requires `Preprocessing`,
-that we offer by means of this very trait.
-
-The Qc library offers several important features
-
-- the definition of an Almanac using
-the [ANISE](https://github.com/nyx-space/anise)
-- the definition of a precise Earth centered Reference [Frame](https://github.com/nyx-space/anise)
-- Injecting [RINEX](https://github.com/georust/rinex) data at the input
-of the pipeline, which is one of the most convenient format for GNSS post processing
-- Injecting [SP3](https://github.com/georust/rinex/tree/main/sp3) data
-at the input of the pipeline, which is the standard format for post processed high precision GNSS navigation.
-- precise timing thanks to the [Hifitime](https://github.com/nyx-space/hifitime) library
-
-## Licensing
-
-These libraries is part of the [RTK-rs framework](https://github.com/rtk-rs) which
+This library is part of the [RTK-rs framework](https://github.com/rtk-rs) which
 is delivered under the [Mozilla V2 Public](https://www.mozilla.org/en-US/MPL/2.0) license.
+
+## Core level
+
+The fundammental blocks that we rely upon, at all times
+
+- [Hifitime by Nyx-Space](https://github.com/nyx-space/hifitime) 
+that provides Epoch and TimeScale definitions
+- [GNSS by RTK-rs](https://github.com/rtk-rs/qc-traits) that provides
+Constellation and SV definitions
+- [Qc Traits by RTK-rs](https://github.com/rtk-rs/qc-traits) that provides 
+shared behavior by all GNSS libraries
+- [The RINEX library by RTK-rs](https://github.com/rtk-rs/rinex) because we consider
+the RINEX files as the most fundamental. It is currently not possible to build
+this library without RINEX support (say: SP3 only application). But that could easily be changed.
+
+## Basic and default features
+
+- `flate2` is activated by default, and allows Gzip compressed files to be naturally supported.
+- `sp3` is activated by default, because we consider people interested in GNSS post processing
+are interested in high precision at all times. This is easily changed by de-activating this crate feature.
+
+## Navigation feature
+
+`nav` is the most advanced feature. It allows post processed navigation and is the heaviest option.
+This option relies on [ANISE by Nyx-Space](https://github.com/nyx-space/anise).
+
+If you are only interested in file processing and management, you should not activate Post Processed navigation support.
+
+## Deploying without navigation support
+
+Without navigation support, this library will allow GNSS context creation and basic processing.
+You will not access the most advanced solvers.
