+++ src/navigation/nav_filter.rs
@@ -64,76 +64,6 @@ impl std::str::FromStr for NavFilter {
     }
 }
 
-use crate::prelude::QcContext;
-
-impl QcContext {
-    /// Applies complex [NavFilter] to mutable [QcContext].
-    pub fn nav_filter_mut(&mut self, filter: &NavFilter) {
-        // apply nav conditions
-        if let Some(brdc) = self.brdc_navigation_mut() {
-            let any_constellation = filter.constellations.is_empty();
-            let broad_sbas = filter.constellations.contains(&Constellation::SBAS);
-
-            let brdc_rec = brdc.record.as_mut_nav().unwrap();
-
-            brdc_rec.retain(|k, data| {
-                if let Some(eph) = data.as_ephemeris() {
-                    match filter.filter {
-                        NavFilterType::Healthy => {
-                            if k.sv.constellation.is_sbas() && broad_sbas {
-                                eph.sv_healthy()
-                            } else {
-                                if any_constellation {
-                                    eph.sv_healthy()
-                                } else {
-                                    if filter.constellations.contains(&k.sv.constellation) {
-                                        eph.sv_healthy()
-                                    } else {
-                                        true
-                                    }
-                                }
-                            }
-                        }
-                        NavFilterType::Testing => {
-                            if k.sv.constellation.is_sbas() && broad_sbas {
-                                eph.sv_in_testing()
-                            } else {
-                                if any_constellation {
-                                    eph.sv_in_testing()
-                                } else {
-                                    if filter.constellations.contains(&k.sv.constellation) {
-                                        eph.sv_in_testing()
-                                    } else {
-                                        true
-                                    }
-                                }
-                            }
-                        }
-                        NavFilterType::Unhealthy => {
-                            if k.sv.constellation.is_sbas() && broad_sbas {
-                                !eph.sv_healthy()
-                            } else {
-                                if any_constellation {
-                                    !eph.sv_healthy()
-                                } else {
-                                    if filter.constellations.contains(&k.sv.constellation) {
-                                        !eph.sv_healthy()
-                                    } else {
-                                        true
-                                    }
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    // preserves other frames
-                    true
-                }
-            });
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::{NavFilter, NavFilterType};
