+++ src/report/mod.rs
@@ -32,8 +32,6 @@ use sp3::SP3Report;
 pub enum Error {
     #[error("non supported RINEX format")]
     NonSupportedRINEX,
-    #[error("sampling analysis failed to return")]
-    SamplingAnalysis,
     #[error("missing Clock RINEX header")]
     MissingClockHeader,
     #[error("missing Meteo RINEX header")]
