+++ src/report/summary/nav_post.rs
@@ -16,10 +16,10 @@ pub struct QcNavPostSummary {
 impl QcNavPostSummary {
     pub fn new(context: &QcContext) -> Self {
         Self {
-            nav_compatible: context.nav_compatible(),
-            cpp_compatible: context.cpp_compatible(),
-            ppp_compatible: context.ppp_compatible(),
-            ppp_ultra_compatible: context.ppp_ultra_compatible(),
+            nav_compatible: context.is_navigation_compatible(),
+            cpp_compatible: context.is_cpp_navigation_compatible(),
+            ppp_compatible: context.is_ppp_navigation_compatible(),
+            ppp_ultra_compatible: context.is_ppp_ultra_navigation_compatible(),
         }
     }
 }
